+++ src/locale.ts
@@ -23,6 +23,17 @@ const messages = {
             loadRawWarning: "A raw json file is loaded",
             overwriteWarning: "Overwrite Warning",
             overwriteDetails: "A history save of current file is detected. This action will OVERWRITE the existed histroy save",
+            switch: {
+                title: "Switch Data Source",
+                explanation: {
+                    raw: "Raw json file",
+                    history: "History save from browser",
+                    server: "Translation by ShinyGroup",
+                    file: "File from disk",
+                },
+                reloadTooltip: "Requery translated file existence",
+                reloadFinished: "Requery finished"
+            }
         },
         home: {
             history: {
@@ -64,6 +75,17 @@ const messages = {
             loadRawWarning: "当前加载源为原始 JSON 文件",
             overwriteWarning: "历史修改将被覆盖",
             overwriteDetails: "检测到浏览器中当前文件的历史修改，本操作会覆盖历史修改，导致历史修改丢失",
+            switch: {
+                title: "切换数据源",
+                explanation: {
+                    raw: "原始JSON文件",
+                    history: "浏览器中的历史修改",
+                    server: "ShinyGroup维护的翻译文件",
+                    file: "磁盘中的文件",
+                },
+                reloadTooltip: "重新查询翻译文件是否存在",
+                reloadFinished: "查询完成"
+            }
         },
         home: {
             history: {